Lexical: Generic type changed to 'T'
Generic type-name placeholder changed to 'T' to follow convention.

This is convention also in C++. The C++ shadow-project (polygons_cxx) used
VTYPE because a form of templates are possible to achieve using macros, and
upper-case names for macro-names is convention in C/C++.

Note that the ambition is to keep the two projects as similar both in design
and implementation as possible so that differences become as clear as
possible (i.e. where respective language is better it should shine and vice
versa).