Linear operations proof on concept
Generics and over-loadable operators IS possible. This is a
proof-of concept. Type gets fairly well restricted as we
have *some* control over the in-arguments, but would be
better if class would be be better restricted to so that T
only can be of scalar type or Vector_n of scalar type.

Improving type-handling is possible but somewhat beyond the
scope (for now).